refactor: (#56) entities의 room 슬라이스 배럴 파일 생성 및 import 경로 수정
