+++ src/entities/room/api/createRoom.ts
@@ -1,5 +1,5 @@
-import { instance } from '@/shared/api/axios';
-import type { Room, CreateRoomPayload } from '@/entities/room/model/types';
+import { instance } from '@/shared';
+import type { CreateRoomPayload, Room } from '../model/types';
 
 export const createRoom = async (data: CreateRoomPayload): Promise<Room> => {
   const response = await instance.post<Room>('rooms', data);
