--- src/entities/room/api/index.ts
@@ -1,2 +1,0 @@
-export * from './getRooms';
-export * from './createRoom';
