+++ src/entities/room/index.ts
@@ -1,2 +1,5 @@
-export * from './api';
-export * from './model';
+export { createRoom } from './api/createRoom';
+export { getRooms } from './api/getRooms';
+export * from './constants/roomConstants';
+export * from './model/types';
+export * from './queries/roomQueries';
